ci: add support to jest testing (#31)
* ci: add support to jest testing

* fix: add support to react-testing-library

issues #30

* fix: remove transform on jest.config.js

need to transform that code into plain JavaScript, similar when building for browsers

* fix: code transformation to javascript

transformed the code to plain Javascript with babel-jest, similar to when building for browsers

issues #30

* fix: add support to jest-expo

added @babel/reset-flow. The preset is recommended to a static type checker for Javascript. Now
jest runs with react and react-na

issues #30

* fix: add test examples and fixed a bug when passing ThemeProvider on native

when running test on native component, jest throws an error about the Consumer.Context. Basically,
ThemeProvider on native should be imported from "styled-components/native" to run correctly

issues #30

Co-authored-by: Mario Souto <soutomarios@gmail.com>